Remove `AivenAclAuthorizer` and Builder
The `AivenAclAuthorizer` is deprecated and unusable with Kafka 3+, the
corresponding `AivenKafkaPrincipalBuilder` is not deprecated but obsolete after
the removal of `AivenAclAuthorizer`. The `AivenAclAuthorizerV2` and its
corresponding `AivenKafkaPrincipalBuilderV2` are the replacements for Kafka 3+,
the `V2` suffix is kept to ensure that the old classes are not used by
accident.

This patch also cleans up the `PrincipalMappers` class that used
`AivenAclAuthorizer` for its log lines, and had a typo.